+++ src/pages/staking-hub/dashboard/safe.tsx
@@ -128,15 +128,15 @@ function SafeDashboard() {
           .finally(() => {
             set_updating(false);
           });
-      } else if (moduleAddress && updateStrategy === 'configWillLetOpenChannels' || updateStrategy === 'configAnnounceOnly') {
+      } else if (moduleAddress && updateStrategy === 'configWillLetOpenChannels') {
+        console.log('WithQ TEST Multi-all')
         // GROUP 2: Safes cloned with old wrong config, but correct SC addresses
 
         const moduleAddressWithout0x = moduleAddress.slice(2).toLocaleLowerCase();
 
-        const newConfig = `0x8d80ff0a0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000016b00${moduleAddressWithout0x}000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000243401cde8000000000000000000000000693bac5ce61c720ddc68533991ceb41199d8f8ae00${moduleAddressWithout0x}000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000243401cde8000000000000000000000000d4fdec44db9d44b8f2b6d529620f9c0c7066a2c100${moduleAddressWithout0x}00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000024a2450f89693bac5ce61c720ddc68533991ceb41199d8f8ae010103030303030303030303000000000000000000000000000000000000000000`;
+        // const newConfig = `0x8d80ff0a0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000016b00${moduleAddressWithout0x}000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000243401cde8000000000000000000000000693bac5ce61c720ddc68533991ceb41199d8f8ae00${moduleAddressWithout0x}000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000243401cde8000000000000000000000000d4fdec44db9d44b8f2b6d529620f9c0c7066a2c100${moduleAddressWithout0x}00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000024a2450f89693bac5ce61c720ddc68533991ceb41199d8f8ae010103030303030303030303000000000000000000000000000000000000000000`;
 
-      //TEST
-       // const newConfig = `0x8d80ff0a0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000016b00${moduleAddressWithout0x}000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000243401cde8000000000000000000000000693bac5ce61c720ddc68533991ceb41199d8f8ae00${moduleAddressWithout0x}000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000243401cde8000000000000000000000000d4fdec44db9d44b8f2b6d529620f9c0c7066a2c100${moduleAddressWithout0x}00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000024a2450f89693bac5ce61c720ddc68533991ceb41199d8f8ae01010303030303030303030300${moduleAddressWithout0x}00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000024a76c9a2f619eabE23FD0E2291B50a507719aa633fE6069b8010003000000000000000000000000000000000000000000000000000000000000`;
+        const newConfig = `0x8d80ff0a000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000001e400${moduleAddressWithout0x}000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000243401cde8000000000000000000000000693bac5ce61c720ddc68533991ceb41199d8f8ae00${moduleAddressWithout0x}000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000243401cde8000000000000000000000000d4fdec44db9d44b8f2b6d529620f9c0c7066a2c100${moduleAddressWithout0x}00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000024a2450f89693bac5ce61c720ddc68533991ceb41199d8f8ae01010303030303030303030300${moduleAddressWithout0x}00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000024a76c9a2f619eabE23FD0E2291B50a507719aa633fE6069b801000300000000000000000000000000000000000000000000000000000000000000000000000000`
 
         dispatch(
           safeActionsAsync.createAndExecuteSafeContractTransactionThunk({
@@ -155,6 +155,7 @@ function SafeDashboard() {
             set_updating(false);
           });
       } if (updateStrategy === 'configAnnounceOnly') {
+        console.log('WithQ configAnnounceOnly')
         // GROUP 3 when announce target was not provided earlier
         const newConfig = `0x619eabE23FD0E2291B50a507719aa633fE6069b8010003000000000000000000`;
 
@@ -233,6 +234,7 @@ function SafeDashboard() {
             set_updating(false);
           });
       } else if (updateStrategy === 'configAnnounceOnly') {
+        console.log('WithQ GROUP 3 ')
         // GROUP 3 when announce target was not provided earlier
         const newConfig = `0x619eabE23FD0E2291B50a507719aa633fE6069b8010003000000000000000000`;
