+++ src/pages/staking-hub/onboarding/index.tsx
@@ -14,20 +14,19 @@ import OnboardingIsFetching from './isFetching';
 import WhatYouWillNeedPage from './step0/0whatYouWillNeed';
 import CreateSafe from './step0/1createSafe';
 import SafeIsReady from './step0/2safeIsReady';
-import ConfigureModule from './step1/0configureModule';
-import OptionalNftTransfer from './step2/0optionalNftTransfer';
-import XdaiToSafe from './step3/0fundSafe';
-import SafeIsFunded from './step3/1safeIsFunded';
-import SelectNodeType from './step4/0selectNodeType';
-import SetupNodeStep from './step4/1setupYourNode';
-import SetupYourDappNode from './step4/1setupYourDappNode';
-import JoinWaitListStep from './step5/0joinWaitlist';
-import AddedToWhitelist from './step5/1addedToWhitelist';
-import AddNode from './step6/0addNode';
-import ConfigureNode from './step7/0configureNode';
-import FundNode from './step8/0fundNode';
-import SetAllowance from './step9/0setAllowance';
-import NodeIsReady from './step10/0nodeIsReady';
+import OptionalNftTransfer from './step1/0optionalNftTransfer';
+import XdaiToSafe from './step2/0fundSafe';
+import SafeIsFunded from './step2/1safeIsFunded';
+import SelectNodeType from './step3/0selectNodeType';
+import SetupNodeStep from './step3/1setupYourNode';
+import SetupYourDappNode from './step3/1setupYourDappNode';
+import JoinWaitListStep from './step4/0joinWaitlist';
+import AddedToWhitelist from './step4/1addedToWhitelist';
+import AddNode from './step5/0addNode';
+import ConfigureNode from './step6/0configureNode';
+import FundNode from './step7/0fundNode';
+import SetAllowance from './step8/0setAllowance';
+import NodeIsReady from './step9/0nodeIsReady';
 
 const OnboardingContainer = styled.div`
   display: flex;
@@ -47,20 +46,19 @@ export const ONBOARDING_PAGES = {
   WHAT_YOU_WILL_NEED: 0,
   CREATE_SAFE: 1,
   SAFE_IS_READY: 2,
-  CONFIGURE_MODULE: 3,
-  OPTIONAL_NFT_TRANSFER: 4,
-  XDAI_TO_SAFE: 5,
-  SAFE_IS_FUNDED: 6,
-  SELECT_NODE_TYPE: 7,
-  SETUP_NODE: 8,
-  SETUP_DAPP_NODE: 9,
+  OPTIONAL_NFT_TRANSFER: 3,
+  XDAI_TO_SAFE: 4,
+  SAFE_IS_FUNDED: 5,
+  SELECT_NODE_TYPE: 6,
+  SETUP_NODE: 7,
+  SETUP_DAPP_NODE: 8,
   JOIN_WAITLIST: 10,
   ADDED_TO_WHITELIST: 11,
   ADD_NODE: 12,
   CONFIGURE_NODE: 13,
   FUND_NODE: 14,
   SET_ALLOWANCE: 15,
-  NODE_IS_READY: 16
+  NODE_IS_READY: 16,
 } as const;
 
 function Onboarding() {
@@ -77,28 +75,26 @@ function Onboarding() {
       case ONBOARDING_PAGES.SAFE_IS_READY:
       case ONBOARDING_PAGES.OPTIONAL_NFT_TRANSFER:
         return 0;
-      case ONBOARDING_PAGES.CONFIGURE_MODULE:
-        return 1;
       case ONBOARDING_PAGES.XDAI_TO_SAFE:
-        return 2;
+        return 1;
       case ONBOARDING_PAGES.SAFE_IS_FUNDED:
       case ONBOARDING_PAGES.SELECT_NODE_TYPE:
       case ONBOARDING_PAGES.SETUP_NODE:
       case ONBOARDING_PAGES.SETUP_DAPP_NODE:
-        return 3;
+        return 2;
       case ONBOARDING_PAGES.JOIN_WAITLIST:
-        return 4;
+        return 3;
       case ONBOARDING_PAGES.ADDED_TO_WHITELIST:
       case ONBOARDING_PAGES.ADD_NODE:
-        return 5;
+        return 4;
       case ONBOARDING_PAGES.CONFIGURE_NODE:
-        return 6;
+        return 5;
       case ONBOARDING_PAGES.FUND_NODE:
-        return 7;
+        return 6;
       case ONBOARDING_PAGES.SET_ALLOWANCE:
-        return 8;
+        return 7;
       case ONBOARDING_PAGES.NODE_IS_READY:
-        return 9;
+        return 8;
 
       default:
         return -1;
@@ -109,29 +105,25 @@ function Onboarding() {
     switch (page) {
       case ONBOARDING_PAGES.OPTIONAL_NFT_TRANSFER:
         return 1;
-      case ONBOARDING_PAGES.CONFIGURE_MODULE:
-        return 2;
       case ONBOARDING_PAGES.XDAI_TO_SAFE:
       case ONBOARDING_PAGES.SAFE_IS_FUNDED:
-        return 3;
+        return 2;
       case ONBOARDING_PAGES.SELECT_NODE_TYPE:
       case ONBOARDING_PAGES.SETUP_NODE:
       case ONBOARDING_PAGES.SETUP_DAPP_NODE:
-        return 4;
+        return 3;
       case ONBOARDING_PAGES.JOIN_WAITLIST:
       case ONBOARDING_PAGES.ADDED_TO_WHITELIST:
-        return 5;
+        return 4;
       case ONBOARDING_PAGES.ADD_NODE:
-        return 6;
+        return 5;
       case ONBOARDING_PAGES.CONFIGURE_NODE:
-        return 7;
-      case ONBOARDING_PAGES.CONFIGURE_MODULE:
-        return 8;
+        return 6;
       case ONBOARDING_PAGES.FUND_NODE:
-        return 9;
+        return 7;
       case ONBOARDING_PAGES.SET_ALLOWANCE:
       case ONBOARDING_PAGES.NODE_IS_READY:
-        return 10;
+        return 8;
 
       default:
         return 0;
@@ -146,7 +138,6 @@ function Onboarding() {
         style={{ flex: '1 1 10%' }}
         steps={[
           { name: 'CREATE SAFE' },
-          { name: 'CONFIGURE MODULE' },
           { name: 'NFT TRANSFER (OPTIONAL)' },
           { name: 'FUND SAFE' },
           { name: 'CHOOSE YOUR NODE SETUP' },
@@ -166,7 +157,6 @@ function Onboarding() {
             {onboardingStep === ONBOARDING_PAGES.WHAT_YOU_WILL_NEED && <WhatYouWillNeedPage />}
             {onboardingStep === ONBOARDING_PAGES.CREATE_SAFE && <CreateSafe />}
             {onboardingStep === ONBOARDING_PAGES.SAFE_IS_READY && <SafeIsReady />}
-            {onboardingStep === ONBOARDING_PAGES.CONFIGURE_MODULE && <ConfigureModule />}
             {onboardingStep === ONBOARDING_PAGES.OPTIONAL_NFT_TRANSFER && <OptionalNftTransfer />}
             {onboardingStep === ONBOARDING_PAGES.XDAI_TO_SAFE && <XdaiToSafe />}
             {onboardingStep === ONBOARDING_PAGES.SAFE_IS_FUNDED && <SafeIsFunded />}
