+++ src/pages/staking-hub/onboarding/nextNode.tsx
@@ -11,10 +11,10 @@ import NetworkOverlay from '../../../components/Overlays/NetworkOverlay';
 
 //Steps
 import OnboardingIsFetching from './isFetching';
-import AddNode from './step6/0addNode';
-import ConfigureNode from './step7/0configureNode';
-import FundNode from './step8/0fundNode';
-import NodeIsReady from './step10/0nodeIsReady';
+import AddNode from './step5/0addNode';
+import ConfigureNode from './step6/0configureNode';
+import FundNode from './step7/0fundNode';
+import NodeIsReady from './step9/0nodeIsReady';
 
 const OnboardingContainer = styled.div`
   display: flex;
