+++ src/pages/staking-hub/onboarding/step1/0optionalNftTransfer.tsx
