+++ src/pages/staking-hub/onboarding/step2/0fundSafe.tsx
