+++ src/pages/staking-hub/onboarding/step2/1safeIsFunded.tsx
