+++ src/pages/staking-hub/onboarding/step4/0joinWaitlist.tsx
