+++ src/pages/staking-hub/onboarding/step4/1addedToWhitelist.tsx
