+++ src/pages/staking-hub/onboarding/step7/0fundNode.tsx
