+++ src/pages/staking-hub/onboarding/step9/0nodeIsReady.tsx
